Separated data initializer stuff from declaration stuff.
